refactor: move classes to subpackages
