--- src/main/java/dinkplugin/SerializedItemStack.java
@@ -1,11 +1,0 @@
-package dinkplugin;
-
-import lombok.Builder;
-
-@Builder
-class SerializedItemStack {
-    private int id;
-    private int quantity;
-    private int priceEach;
-    private String name;
-}
