+++ src/main/java/dinkplugin/notifiers/CollectionNotifier.java
@@ -1,4 +1,10 @@
-package dinkplugin;
+package dinkplugin.notifiers;
+
+import dinkplugin.DinkPlugin;
+import dinkplugin.message.NotificationBody;
+import dinkplugin.message.NotificationType;
+import dinkplugin.Utils;
+import dinkplugin.notifiers.data.CollectionNotificationData;
 
 import javax.inject.Inject;
