+++ src/main/java/dinkplugin/notifiers/LootNotifier.java
@@ -1,5 +1,11 @@
-package dinkplugin;
+package dinkplugin.notifiers;
 
+import dinkplugin.DinkPlugin;
+import dinkplugin.message.NotificationBody;
+import dinkplugin.message.NotificationType;
+import dinkplugin.Utils;
+import dinkplugin.notifiers.data.LootNotificationData;
+import dinkplugin.notifiers.data.SerializedItemStack;
 import net.runelite.api.ItemComposition;
 import net.runelite.client.game.ItemStack;
 import net.runelite.client.util.QuantityFormatter;
