+++ src/main/java/dinkplugin/notifiers/PetNotifier.java
@@ -1,4 +1,9 @@
-package dinkplugin;
+package dinkplugin.notifiers;
+
+import dinkplugin.DinkPlugin;
+import dinkplugin.message.NotificationBody;
+import dinkplugin.message.NotificationType;
+import dinkplugin.Utils;
 
 import javax.inject.Inject;
