+++ src/main/java/dinkplugin/notifiers/QuestNotifier.java
@@ -1,4 +1,10 @@
-package dinkplugin;
+package dinkplugin.notifiers;
+
+import dinkplugin.DinkPlugin;
+import dinkplugin.message.NotificationBody;
+import dinkplugin.message.NotificationType;
+import dinkplugin.Utils;
+import dinkplugin.notifiers.data.QuestNotificationData;
 
 public class QuestNotifier extends BaseNotifier {
