+++ src/main/java/dinkplugin/notifiers/SlayerNotifier.java
@@ -1,5 +1,10 @@
-package dinkplugin;
+package dinkplugin.notifiers;
 
+import dinkplugin.DinkPlugin;
+import dinkplugin.message.NotificationBody;
+import dinkplugin.message.NotificationType;
+import dinkplugin.Utils;
+import dinkplugin.notifiers.data.SlayerNotificationData;
 import lombok.Getter;
 import lombok.Setter;
