+++ src/main/java/dinkplugin/notifiers/data/SerializedItemStack.java
@@ -1,0 +1,15 @@
+package dinkplugin.notifiers.data;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+@AllArgsConstructor
+public class SerializedItemStack {
+    private int id;
+    private int quantity;
+    private int priceEach;
+    private String name;
+}
