+++ src/main/java/dinkplugin/notifiers/data/SpeedrunNotificationData.java
@@ -1,8 +1,10 @@
-package dinkplugin;
+package dinkplugin.notifiers.data;
 
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 
 @Data
+@EqualsAndHashCode(callSuper = true)
 public class SpeedrunNotificationData extends QuestNotificationData {
     private String personalBest;
     private String currentTime;
