+++ src/test/java/dinkplugin/IgnoredWorldTest.kt
@@ -6,7 +6,6 @@ import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.ArgumentsProvider
 import org.junit.jupiter.params.provider.ArgumentsSource
-import org.junit.jupiter.params.provider.ValueSource
 import java.util.stream.Stream
 import java.util.EnumSet;
 import net.runelite.api.WorldType;
@@ -15,13 +14,13 @@ class IgnoredWorldTest {
     @ParameterizedTest(name = "World should not be ignored {0}")
     @ArgumentsSource(NonIgnoredWorldTypeProvider::class)
     fun `World should not be ignored`(worldType: Set<WorldType>) {
-        assertFalse(DinkPlugin._isIgnoredWorld(worldType))
+        assertFalse(Utils.isIgnoredWorld(worldType))
     }
 
     private class NonIgnoredWorldTypeProvider : ArgumentsProvider {
         override fun provideArguments(context: ExtensionContext?): Stream<out Arguments> = Stream.of(
             Arguments.of(EnumSet.of(WorldType.MEMBERS)),
-            Arguments.of(emptySet<WorldType>()),
+            Arguments.of(EnumSet.noneOf(WorldType::class.java)),
             Arguments.of(EnumSet.of(WorldType.SKILL_TOTAL)),
             Arguments.of(EnumSet.of(WorldType.FRESH_START_WORLD)),
         )
@@ -30,7 +29,7 @@ class IgnoredWorldTest {
     @ParameterizedTest(name = "World should be ignored {0}")
     @ArgumentsSource(IgnoredWorldTypeProvider::class)
     fun `World should be ignored`(worldType: Set<WorldType>) {
-        assertTrue(DinkPlugin._isIgnoredWorld(worldType))
+        assertTrue(Utils.isIgnoredWorld(worldType))
     }
 
     private class IgnoredWorldTypeProvider : ArgumentsProvider {
